refactor(portfolio): dedup P&L/dividend aggregation + extract report-schedule service
Two fixes to routers/portfolio.py (934 lines):

1. DRY: realized-P&L and dividend aggregation ("filter SELL txns / dividends in
   a period, FX-convert, accumulate") was copy-pasted in get_portfolio_summary
   and the report-schedule handler. Extract PortfolioService.get_realized_pnl_by_asset
   and get_dividend_income_by_asset (raising FxRateError; the router maps to 422).
   Both endpoints now share them — duplication and drift gone.

2. Layering: get_investment_performance_report_schedule was a 341-line handler
   (36% of the router) inlining holdings/allocation/freshness/aggregation domain
   logic. Move it to services/performance_report.build_investment_performance_report_schedule;
   the route is now thin validation + delegation. The report-only helpers
   (_money/_percent/_source_document_links/_freshness_link/_report_preparation_holdings)
   move with it.

A shared `portfolio_service` singleton (services/portfolio.py) is now imported by
both the router and the new service so monkeypatching one reference affects both.

routers/portfolio.py: 934 -> 538 lines. Behavior unchanged; routes keep paths/
signatures (reports.py still calls the thin route fn). ruff clean; 138 portfolio
+ 275 reporting/package tests green.
