+++ apps/backend/tests/portfolio/test_financial_logic_audit.py
@@ -10,7 +10,7 @@ from src.models import Account, AccountType, FxRate
 from src.models.layer2 import AtomicPosition, AtomicTransaction, TransactionDirection
 from src.models.layer3 import CostBasisMethod, ManagedPosition, PositionStatus
 from src.models.portfolio import DividendIncome, InvestmentTransaction, InvestmentTransactionType
-from src.routers.portfolio import _source_document_links
+from src.services.performance_report import _source_document_links
 from src.services.performance import calculate_money_weighted_return, calculate_time_weighted_return, calculate_xirr
